requirements.txt
fastapi

uvicorn

pydantic

google-genai

asyncio

multiprocessing